Enable -Wunderspecs in Makefile.
Underspecs ensure that dialyzer will warn about types that are underspecified:
the specification allows a wider set of types than dialyzer can infer. Enabling
this _also_ seems to catch errors that are not set width so much as just not
applicable, [[spec_fun()]] for the desired [spec_fun()] in the last commit.

Had -Wunderspecs been enabled the last bug would not have gone through.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>